+++ app/code/local/Ffuenf/UpdateScripts/data/ffuenf_updatescripts_setup/data-install-1.0.0.php
@@ -34,70 +34,56 @@ foreach ($variableNames as $variableName) {
 */
 
 /*
- * add transactional email templates
+ * add transactional email template
  *
-$localeEmailPath = Mage::helper('ffuenf_updatescripts')->getLocaleEmailPath();
-$templateCode = 'YOUR TEMPLATE CODE';
-$template = Mage::getModel('core/email_template')->loadByCode($templateCode);
-$template->setTemplateCode($templateCode)
-         ->setTemplateType(2)
-         ->setAddedAt(Mage::getSingleton('core/date')->gmtDate())
-         ->setModifiedAt(Mage::getSingleton('core/date')->gmtDate());
-// load template content from file
-$templateText = Mage::helper('ffuenf_updatescripts')->getTemplateContent($localeEmailPath . 'sales/sometemplate.html');
-if (!$templateText) {
-    Mage::log('Failed to create new E-Mail Template: ' . $templateCode);
-    return;
-}
-if (preg_match('/<!--@subject\s*(.*?)\s*@-->/u', $templateText, $matches)) {
-    $template->setTemplateSubject($matches[1]);
-    $templateText = str_replace($matches[0], '', $templateText);
-}
-if (preg_match('/<!--@vars\s*((?:.)*?)\s*@-->/us', $templateText, $matches)) {
-    $templateText = str_replace($matches[0], '', $templateText);
-}
-if (preg_match('/<!--@styles\s*(.*?)\s*@-->/s', $templateText, $matches)) {
-    $template->setTemplateStyles($matches[1]);
-    $templateText = str_replace($matches[0], '', $templateText);
-}
-$templateText = preg_replace('#\{\*.*\*\}#suU', '', $templateText);
-$template->setTemplateText($templateText)->save();
-Mage::log('Created new E-Mail Template: ' . $templateCode);
+$identifier      = '_identifier';
+$storeId         = 1;
+$locale          = 'en_US';
+$templateContent = Mage::helper('ffuenf_updatescripts')->getTemplateContent($identifier, $locale, 'email');
+$cmsPage         = array(
+                    'title'      => 'Your title',
+                    'identifier' => $identifier,
+                    'content'    => $templateContent
+                 );
+Mage::helper('ffuenf_updatescripts')->saveCmsData($cmsPage, $storeId, 'email');
 */
 
 /*
  * add cms static block
  *
-$storeId = 1;
-$cmsBlock = array(
-    'title' => 'Your title',
-    'identifier' => 'identifier',
-    'content' => 'your content',
-    'is_active' => 1,
-    'sort_order' => 0,
-    'stores' => array($storeId),
-    'root_template' => 'one_column',
-    'view_mode' => 'core' # or 'easytemplate' for Webguys_EasyTemplate
-);
-Mage::helper('ffuenf_updatescripts')->saveCmsData($cmsBlock, $storeId, false);
-Mage::log('Created new CMS block: ' . $cmsBlock['title']);
+$identifier      = '_identifier';
+$storeId         = 1;
+$locale          = 'en_US';
+$templateContent = Mage::helper('ffuenf_updatescripts')->getTemplateContent($identifier, $locale, 'block');
+$cmsBlock        = array(
+                    'title'      => 'Your title',
+                    'identifier' => $identifier,
+                    'content'    => $templateContent,
+                    'is_active'  => 1,
+                    'sort_order' => 0,
+                    'stores'     => array($storeId),
+                    'view_mode'  => 'core' # or 'easytemplate' for Webguys_EasyTemplate
+                 );
+Mage::helper('ffuenf_updatescripts')->saveCmsData($cmsBlock, $storeId, 'block');
 */
 
 /*
  * add cms page
  *
-$storeId = 1;
-$cmsPage = array(
-    'title' => 'Your title',
-    'identifier' => 'identifier',
-    'content' => 'your content',
-    'is_active' => 1,
-    'sort_order' => 0,
-    'stores' => array($storeId),
-    'view_mode' => 'core' # or 'easytemplate' for Webguys_EasyTemplate
-);
-Mage::helper('ffuenf_updatescripts')->saveCmsData($cmsPage, $storeId, true);
-Mage::log('Created new CMS page: ' . $cmsPage['title']);
+$identifier      = '_identifier';
+$storeId         = 1;
+$locale          = 'en_US';
+$templateContent = Mage::helper('ffuenf_updatescripts')->getTemplateContent($identifier, $locale, 'page');
+$cmsPage         = array(
+                    'title'      => 'Your title',
+                    'identifier' => $identifier,
+                    'content'    => $templateContent,
+                    'is_active'  => 1,
+                    'sort_order' => 0,
+                    'stores'     => array($storeId),
+                    'view_mode'  => 'core' # or 'easytemplate' for Webguys_EasyTemplate
+                 );
+Mage::helper('ffuenf_updatescripts')->saveCmsData($cmsPage, $storeId, 'page');
 */
 
 $installer->endSetup();